various
- system devices as member vars
- integer player
- file renames

+++ app_player_adapter.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "algorithm.hpp"
+#include "alg_ringindex.hpp"
 #include "kb_tinymod.hpp"
 
 #include <cstdint>
@@ -40,7 +40,7 @@ private:
 private:
 	kb::ModPlayer& player_;
 	alg::RingIndex<4096> rw_;
-	std::int16_t buf_[4096*2];
+	std::int32_t buf_[4096*2];
 	int pbuf_[4096*2]; };
